From this commit all external libraries should be in "vendor" directory. The vendor directory should be:
- app/resources/libs/vendor
- any vendor directory inside a plugin, module...

Prettier ignore all js and css files with extension .min.js or .min.css,
in fact all external libraries should be in minimified format.

I'm migrating slowly, but very soon icestudio use a bundler that
improve the developing experience, this bundlers minimify all the needed
files.

All .ice, .md , .txt files are ignored too

In my side left write a DEVELOPER.md with clear guidelines of how to
develop, as soon as i close some task in progress that modify a lot the
way of develop in Icestudio, i'll document it to help current and future
collaborators.

Thanks a lot for all of you that are collaborating in icestudio, we
are buildiing the better EDA editor!